+++ 2025-Reefscape/src/main/include/subsystems/Elevator.h
@@ -30,7 +30,7 @@ class Elevator : public frc2::SubsystemBase {
 
     double getPosition();
     double getRelativePosition();
-    void setSpeed(int speed);
+    void setSpeed(double speed);
     void resetMotors();
     void resetEncoders();
     void levelOne();
